sdk/runner/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
prometheus-client>=0.11.0
opentelemetry-api>=1.12.0
opentelemetry-sdk>=1.12.0
requests>=2.25.0
